+++ apps/gizual-app/src/primitives/file-tree/file-tree.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable unicorn/no-null */
 import { IconCollapse } from "@app/assets";
-import { useViewModel } from "@app/services/view-model";
 import { Checkbox } from "@mantine/core";
 import { notifications } from "@mantine/notifications";
 import clsx from "clsx";
@@ -33,7 +32,11 @@ const MAX_RENDER_DEPTH = 3;
 
 export const FileTree = observer(({ files, checked, onChange }: FileTreeProps) => {
   const availableFiles = files ?? useAvailableFiles();
-  const vm = useViewModel(FileTreeViewModel, availableFiles, checked);
+  if (!availableFiles) return <>Empty set of files.</>;
+
+  const vm = React.useMemo(() => {
+    return new FileTreeViewModel(availableFiles, checked);
+  }, [availableFiles]);
 
   React.useEffect(() => {
     vm.assignCheckedState(checked);