+++ apps/gizual-app/src/primitives/file-tree/file-tree.vm.tsx
@@ -1,5 +1,4 @@
-import { Dependencies, ViewModel } from "@app/services/view-model";
-import { action, computed, makeObservable, observable, toJS } from "mobx";
+import { action, computed, makeObservable, observable } from "mobx";
 import { match } from "ts-pattern";
 
 export type FileTreeMode = "full" | "favorites";
@@ -23,7 +22,7 @@ export type FileTreeNode = {
   parentPath: string[];
 };
 
-export class FileTreeViewModel extends ViewModel {
+export class FileTreeViewModel {
   /* List of all available files we could render. */
   private _availableFiles!: FileTreeFlatItem[];
 
@@ -36,30 +35,11 @@ export class FileTreeViewModel extends ViewModel {
   /* Root of the constructed tree. */
   private _root: FileTreeNode | undefined;
 
-  constructor({ mainController }: Dependencies, ...args: any[]) {
-    super({ mainController }, ...args);
-    makeObservable(this, undefined);
-  }
-
-  init(...args: any[]): void {
-    super.init(...args);
-
-    const availableFiles = toJS(args[0]) as FileTreeFlatItem[];
-    const checked = toJS(args[1]) as string[];
-
-    if (!availableFiles || availableFiles.length === 0) {
-      return;
-    }
-
-    this._availableFiles = availableFiles
-      .sort((a, b) => {
-        if (a.kind === "folder" && b.kind !== "folder") return -1;
-        return a.path.length - b.path.length;
-      })
-      .filter((f) => f.path.length > 0);
-
+  constructor(availableFiles: FileTreeFlatItem[], checked: string[] | undefined) {
+    this._availableFiles = availableFiles;
     this.constructTree();
     if (checked) this.assignCheckedState(checked);
+    makeObservable(this, undefined, { autoBind: true });
   }
 
   /**
@@ -97,7 +77,6 @@ export class FileTreeViewModel extends ViewModel {
       });
 
       this._nodes[path.join("/")] = node;
-      if (!this._nodes[parentPath]) this._nodes[parentPath] = { ...root, path: path.slice(0, -1) };
       this._nodes[parentPath].children.push(node);
     }