+++ apps/gizual-app/src/primitives/gradient-legend/gradient-legend.module.scss.d.ts
@@ -1,5 +1,6 @@
 declare const classNames: {
   readonly GradientLegend: "GradientLegend";
   readonly GradientTextElement: "GradientTextElement";
+  readonly GradientLegendActions: "GradientLegendActions";
 };
 export = classNames;