+++ packages/explorer-web/ts-src/pool/pool-node.ts
@@ -1,5 +1,6 @@
 import { isString } from "lodash";
 
+import { FSHandle } from "@giz/opfs";
 import { WasiRuntime } from "@giz/wasi-runtime";
 import wasmFileUrl from "../../build/explorer-web.wasm?url";
 
@@ -38,11 +39,11 @@ export class PoolNode {
     }
   }
 
-  boot(handle: FileSystemDirectoryHandle | Uint8Array) {
+  boot(handle: FSHandle | Uint8Array) {
     this.starting = this._boot(handle);
   }
 
-  private async _boot(handle: FileSystemDirectoryHandle | Uint8Array): Promise<void> {
+  private async _boot(handle: FSHandle | Uint8Array): Promise<void> {
     const runtime = await WasiRuntime.create({
       moduleUrl: this.wasmFileUrl,
       moduleName: "/module.wasm",