+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **app:general:** Implement custom loggers with pretty prefixes, namespaces and custom filtering
 - **app:general:** Allow direct remote cloning via query parameter `source`
+- **app:backend:** Implement initial support for safari by executing remote-clone in a web-worker
 
 ### Changes